Move other App tests into newer App.test file
